Move plotting code to plots module
